# Conflicts:
#	Forecast/app/src/main/AndroidManifest.xml
#	Forecast/app/src/main/java/com/khasang/forecast/MyApplication.java

+++ Forecast/app/src/main/java/com/khasang/forecast/activities/CityPickerActivity.java
@@ -13,7 +13,6 @@ import android.support.design.widget.Snackbar;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.app.Fragment;
 import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
@@ -26,7 +25,6 @@ import android.view.KeyEvent;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.view.ViewGroup;
 import android.view.WindowManager;
 import android.view.animation.AccelerateInterpolator;
 import android.view.animation.Animation;
@@ -40,13 +38,13 @@ import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.khasang.forecast.AppUtils;
 import com.khasang.forecast.Maps;
-import com.khasang.forecast.MyApplication;
 import com.khasang.forecast.R;
 import com.khasang.forecast.adapters.CityPickerAdapter;
 import com.khasang.forecast.adapters.GooglePlacesAutocompleteAdapter;
 import com.khasang.forecast.adapters.etc.HidingScrollListener;
+import com.khasang.forecast.api.GoogleMapsGeocoding;
+import com.khasang.forecast.api.GoogleMapsTimezone;
 import com.khasang.forecast.exceptions.EmptyCurrentAddressException;
 import com.khasang.forecast.exceptions.NoAvailableAddressesException;
 import com.khasang.forecast.interfaces.IMapDataReceiver;
@@ -54,6 +52,7 @@ import com.khasang.forecast.interfaces.IMessageProvider;
 import com.khasang.forecast.location.LocationParser;
 import com.khasang.forecast.position.Coordinate;
 import com.khasang.forecast.position.PositionManager;
+import com.khasang.forecast.utils.AppUtils;
 import com.khasang.forecast.view.DelayedAutoCompleteTextView;
 import com.mikepenz.community_material_typeface_library.CommunityMaterial;
 import com.mikepenz.google_material_typeface_library.GoogleMaterial;
@@ -67,44 +66,51 @@ import java.util.Locale;
 import java.util.Set;
 import java.util.regex.Pattern;
 
+import butterknife.BindView;
+
 /**
  * Created by CopyPasteStd on 29.11.15.
- * <p/>
+ * <p>
  * Activity для выбора местоположения
  */
+public class CityPickerActivity extends BaseActivity implements View.OnClickListener, IMapDataReceiver,
+        IMessageProvider {
 
-public class CityPickerActivity extends AppCompatActivity implements View.OnClickListener, IMapDataReceiver, IMessageProvider {
-    private final String TAG = this.getClass().getSimpleName();
     public final static String CITY_PICKER_TAG = "com.khasang.forecast.activities.CityPickerActivity";
+    private final String TAG = this.getClass().getSimpleName();
 
-    private TextView infoTV;
-    private RecyclerView recyclerView;
+    @BindView(R.id.infoTV) TextView infoTV;
+    @BindView(R.id.recyclerView) RecyclerView recyclerView;
+    @BindView(R.id.toolbar) Toolbar toolbar;
+    @BindView(R.id.fabBtn) volatile FloatingActionButton fabBtn;
+
     private CityPickerAdapter cityPickerAdapter;
-    List<String> cityList;
-
-    private Toolbar toolbar;
-    private volatile FloatingActionButton fabBtn;
+    private List<String> cityList;
 
     private DelayedAutoCompleteTextView chooseCity;
+    GoogleMapsGeocoding googleMapsGeocoding;
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setTheme(AppUtils.getCurrentTheme(this));
         setContentView(R.layout.activity_city_picker);
-        toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-        if (getSupportActionBar() != null) {
-            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        }
-        setTitle(getString(R.string.city_list));
 
-        infoTV = (TextView) findViewById(R.id.infoTV);
-        recyclerView = (RecyclerView) findViewById(R.id.recyclerView);
+        setSupportActionBar(toolbar);
+        //noinspection ConstantConditions
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        getSupportActionBar().setTitle(R.string.title_activity_city_picker);
+        toolbar.setNavigationOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                leaveActivity();
+            }
+        });
+
         cityList = new ArrayList<>();
         cityPickerAdapter = new CityPickerAdapter(cityList, this);
         recyclerView.setAdapter(cityPickerAdapter);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
+        googleMapsGeocoding = new GoogleMapsGeocoding();
 
         swapVisibilityTextOrList();
 
@@ -120,7 +126,6 @@ public class CityPickerActivity extends AppCompatActivity implements View.OnClic
                 showViews();
             }
         });
-        fabBtn = (FloatingActionButton) findViewById(R.id.fabBtn);
         IconicsDrawable icon = new IconicsDrawable(this)
                 .color(ContextCompat.getColor(this, R.color.current_weather_color))
                 .icon(CommunityMaterial.Icon.cmd_plus);
@@ -132,9 +137,11 @@ public class CityPickerActivity extends AppCompatActivity implements View.OnClic
         setupFooterHeight();
         createItemList();
 
-        ItemTouchHelper.SimpleCallback simpleCallback = new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT) {
+        ItemTouchHelper.SimpleCallback simpleCallback = new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT |
+                ItemTouchHelper.RIGHT) {
             @Override
-            public boolean onMove(RecyclerView recyclerView, RecyclerView.ViewHolder viewHolder, RecyclerView.ViewHolder target) {
+            public boolean onMove(RecyclerView recyclerView, RecyclerView.ViewHolder viewHolder, RecyclerView
+                    .ViewHolder target) {
                 return false;
             }
 
@@ -291,10 +298,8 @@ public class CityPickerActivity extends AppCompatActivity implements View.OnClic
     }
 
     private Coordinate getTownCoordinates(String city) {
-
         Coordinate coordinate;
         if (city.length() <= 0) {
-            showToast(R.string.error_empty_location_name);
             return null;
         }
         Geocoder geocoder = new Geocoder(getApplicationContext());
@@ -302,7 +307,6 @@ public class CityPickerActivity extends AppCompatActivity implements View.OnClic
         try {
             addresses = geocoder.getFromLocationName(city, 3);
             if (addresses.size() == 0) {
-                showToast(R.string.coordinates_not_found);
                 return null;
             }
             Address currentAddress = addresses.get(0);
@@ -310,11 +314,12 @@ public class CityPickerActivity extends AppCompatActivity implements View.OnClic
             coordinate.setLatitude(currentAddress.getLatitude());
             coordinate.setLongitude(currentAddress.getLongitude());
         } catch (IOException e) {
-            showToast(R.string.error_geo_service_not_available);
             e.printStackTrace();
             return null;
         } catch (IllegalArgumentException | NullPointerException e) {
-            showToast(R.string.invalid_lang_long_used);
+            e.printStackTrace();
+            return null;
+        } catch (Exception e) {
             e.printStackTrace();
             return null;
         }
@@ -323,8 +328,10 @@ public class CityPickerActivity extends AppCompatActivity implements View.OnClic
 
     // Вспомогательный метод для добавления города в список
     private void addItem(String city, Coordinate coordinate) {
+        boolean needCoordinateRequest = false;
         if (coordinate == null) {
             coordinate = new Coordinate(0, 0);
+            needCoordinateRequest = true;
         }
         if (!PositionManager.getInstance().positionInListPresent(city)) {
             PositionManager.getInstance().addPosition(city, coordinate);
@@ -332,8 +339,14 @@ public class CityPickerActivity extends AppCompatActivity implements View.OnClic
             cityList.add(city);
             Collections.sort(cityList);
             cityPickerAdapter.notifyDataSetChanged();
+            if (needCoordinateRequest) {
+                googleMapsGeocoding.requestCoordinates(city, PositionManager.getInstance(), true);
+            } else {
+                GoogleMapsTimezone googleMapsTimezone = new GoogleMapsTimezone();
+                googleMapsTimezone.requestCoordinates(city);
+            }
         } else {
-            showMessageToUser(R.string.city_exist, Snackbar.LENGTH_LONG);
+            showSnackbar(R.string.city_exist, Snackbar.LENGTH_LONG);
         }
         swapVisibilityTextOrList();
     }
@@ -348,38 +361,33 @@ public class CityPickerActivity extends AppCompatActivity implements View.OnClic
         Fragment frag = getSupportFragmentManager().findFragmentById(R.id.map);
         map.closeDataReceiver();
         if (frag != null) {
-            getSupportFragmentManager().beginTransaction().remove(getSupportFragmentManager().findFragmentById(R.id.map)).commit();
+            getSupportFragmentManager().beginTransaction().remove(getSupportFragmentManager().findFragmentById(R.id
+                    .map)).commit();
         }
     }
 
-    private void setLocationOnMap(Maps maps, String city) {
+    private void setLocationOnMap(final Maps maps, final String city) {
         try {
             Coordinate coordinate = getTownCoordinates(city);
-            double latitude = coordinate != null ? coordinate.getLatitude() : 0;
-            double longitude = coordinate != null ? coordinate.getLongitude() : 0;
-
-            maps.deleteAllMarkers();
-            maps.setNewMarker(latitude, longitude, city);
-            maps.setCameraPosition(latitude, longitude, maps.getDefaultZoom(), 0, 0);
+            if (coordinate != null) {
+                maps.deleteAllMarkers();
+                maps.setNewMarker(coordinate, city);
+                maps.setCameraPosition(coordinate, maps.getDefaultZoom(), 0, 0);
+            } else {
+                maps.setNewMarker(city);
+            }
         } catch (Exception e) {
             e.printStackTrace();
         }
     }
 
-    public String ConvertPointToLocation(double latitude, double longitude) {
+    public String convertPointToLocation(double latitude, double longitude) {
         String address = "";
         Geocoder geoCoder = new Geocoder(getBaseContext(), Locale.getDefault());
         try {
             List<Address> addresses = geoCoder.getFromLocation(latitude, longitude, 3);
             address = new LocationParser(addresses).parseList().getAddressLine();
-        } catch (IOException e) {
-            showToast(R.string.error_service_not_available);
-            e.printStackTrace();
-        } catch (IllegalArgumentException e) {
-            showToast(R.string.invalid_lang_long_used);
-            e.printStackTrace();
-        } catch (EmptyCurrentAddressException | NoAvailableAddressesException e) {
-            showToast(R.string.no_address_found);
+        } catch (IOException | IllegalArgumentException | EmptyCurrentAddressException | NoAvailableAddressesException e) {
             e.printStackTrace();
         } catch (Exception e) {
             e.printStackTrace();
@@ -394,7 +402,7 @@ public class CityPickerActivity extends AppCompatActivity implements View.OnClic
 
     @Override
     public String setLocationCoordinatesFromMap(double latitude, double longitude) throws EmptyCurrentAddressException {
-        String location = ConvertPointToLocation(latitude, longitude);
+        String location = convertPointToLocation(latitude, longitude);
         if (location.isEmpty()) {
             chooseCity.setText("");
             throw new EmptyCurrentAddressException();
@@ -403,6 +411,14 @@ public class CityPickerActivity extends AppCompatActivity implements View.OnClic
         return location;
     }
 
+    @Override
+    public void setLocation(String city) {
+        if (city == null) {
+            showToast(R.string.no_address_found);
+        } else {
+            chooseCity.setText(city);
+        }
+    }
 
     private void setBtnClear(View view) {
         view.findViewById(R.id.btnClear).setOnClickListener(new View.OnClickListener() {
@@ -422,12 +438,11 @@ public class CityPickerActivity extends AppCompatActivity implements View.OnClic
                 try {
                     Coordinate coordinate = getTownCoordinates(chooseCity.getAdapter().getItem(i).toString());
                     if (coordinate == null) {
-                        showToast(R.string.invalid_lang_long_used);
                         return;
                     }
-                    maps.setNewMarker(coordinate.getLatitude(), coordinate.getLongitude(), chooseCity.getAdapter().getItem(i).toString());
+                    maps.setNewMarker(coordinate.getLatitude(), coordinate.getLongitude(), chooseCity.getAdapter()
+                            .getItem(i).toString());
                 } catch (NullPointerException e) {
-                    showToast(R.string.invalid_lang_long_used);
                     e.printStackTrace();
                 } catch (Exception e) {
                     e.printStackTrace();
@@ -439,17 +454,20 @@ public class CityPickerActivity extends AppCompatActivity implements View.OnClic
     // Метод скрытия клавиатуры
     public void hideSoftKeyboard(Context context) {
         if (chooseCity.isFocused()) {
-            InputMethodManager inputMethodManager = (InputMethodManager) context.getSystemService(Context.INPUT_METHOD_SERVICE);
+            InputMethodManager inputMethodManager = (InputMethodManager) context.getSystemService(Context
+                    .INPUT_METHOD_SERVICE);
             inputMethodManager.toggleSoftInput(InputMethodManager.SHOW_FORCED, 0);
         }
     }
 
     private void showChooseCityDialog() {
         final Pattern pattern = Pattern.compile("^[\\w\\s,`'()-]+$");
-        final View view = getLayoutInflater().inflate(R.layout.dialog_pick_location, (ViewGroup) null);
+        final View view = getLayoutInflater().inflate(R.layout.dialog_pick_location, null);
+
         AlertDialog.Builder builder = new AlertDialog.Builder(this);
         setBtnClear(view);
-        final GooglePlacesAutocompleteAdapter googlePlacesAutocompleteAdapter = new GooglePlacesAutocompleteAdapter(this, R.layout.autocomplete_city_textview_item);
+        final GooglePlacesAutocompleteAdapter googlePlacesAutocompleteAdapter = new GooglePlacesAutocompleteAdapter
+                (this, R.layout.autocomplete_city_textview_item);
         final Maps map = new Maps(this, this, this);
         chooseCity = (DelayedAutoCompleteTextView) view.findViewById(R.id.editTextCityName);
         chooseCity.setThreshold(3);
@@ -531,7 +549,8 @@ public class CityPickerActivity extends AppCompatActivity implements View.OnClic
                     return;
                 }
                 char lastSym = s.toString().toCharArray()[s.length() - 1];
-                if ((chooseCity.getText().toString().trim().length() % 4 == 0) || (lastSym == ' ') || (lastSym == '-')) {
+                if ((chooseCity.getText().toString().trim().length() % 4 == 0) || (lastSym == ' ') || (lastSym ==
+                        '-')) {
                     setMarkersOnMap(map);
                 }
                 if (s.toString().contains("\n")) {
@@ -588,19 +607,13 @@ public class CityPickerActivity extends AppCompatActivity implements View.OnClic
     }
 
     @Override
-    public boolean onSupportNavigateUp() {
-        onBackPressed();
-        return true;
+    public void showSnackbar(int stringId, int length) {
+        showSnackbar(getString(stringId), length);
     }
 
     @Override
-    public void showMessageToUser(CharSequence string, int length) {
+    public void showSnackbar(CharSequence string, int length) {
         AppUtils.showSnackBar(this, findViewById(R.id.fabBtn), string, length);
-    }
-
-    @Override
-    public void showMessageToUser(int stringId, int length) {
-        showMessageToUser(getString(stringId), length);
     }
 
     @Override
@@ -611,7 +624,7 @@ public class CityPickerActivity extends AppCompatActivity implements View.OnClic
     @Override
     public void showToast(CharSequence string) {
         Toast toast = AppUtils.showInfoMessage(this, string);
-        toast.getView().setBackgroundColor(ContextCompat.getColor(MyApplication.getAppContext(), R.color.background_toast));
+        toast.getView().setBackgroundColor(ContextCompat.getColor(this, R.color.background_toast));
         toast.setGravity(Gravity.CENTER, 0, 0);
         toast.show();
     }
